refactor: clean up quadrotor certificate files
Remove 484 lines of commented-out dead code from barrier_functions.py
(686 -> 202 lines). Split lyapunov_functions.py (716 lines) into a
package with basic.py (position/attitude/velocity) and advanced.py
(composite/position-velocity/geometric).
